progetto-funzionante-master/progetto-funzionante-master/frontend/requirements.txt
# Core FastAPI - compatible with Railway
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0

# Auth
passlib>=1.7.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.20
pydantic>=2.0.0,<3.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0

# HTTP - fundamental packages
httpx>=0.25.0,<0.30.0
requests>=2.28.0,<3.0.0

# Data processing for OANDA and technical analysis
pandas>=1.5.0,<2.3.0
numpy>=1.21.0,<2.0.0

# OANDA API integration
aiohttp>=3.8.0,<4.0.0

# AI integration (Gemini)
google-generativeai>=0.3.0

# ML System dependencies - Required for Quant Adaptive System
aiofiles>=0.8.0,<1.0.0
aiosqlite>=0.18.0,<1.0.0
scikit-learn>=1.3.0,<2.0.0
beautifulsoup4>=4.11.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Redis caching for performance optimization
redis>=4.5.0,<6.0.0
aioredis>=2.0.0,<3.0.0

# Advanced caching and performance
cachetools>=5.3.0,<6.0.0
cryptography>=41.0.0,<42.0.0

# Sentry error tracking and monitoring
sentry-sdk>=1.39.0,<2.0.0

# Advanced metrics and monitoring
prometheus-client>=0.19.0,<1.0.0

# Logging and monitoring
structlog>=23.1.0,<24.0.0
python-json-logger>=2.0.0,<3.0.0
loguru>=0.7.0,<1.0.0
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.41b0,<1.0.0
opentelemetry-exporter-jaeger>=1.20.0,<2.0.0
opentelemetry-exporter-prometheus>=1.20.0,<2.0.0
psutil>=5.9.0,<6.0.0
watchdog>=3.0.0,<4.0.0

# Alerting and notifications
slack-sdk>=3.21.0,<4.0.0

# Analytics and reporting
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0
